Instead of extracting URLs wherever it sees them, follow a fixed course around the API.
Strips out part of the script that finds new URLs everywhere, and instead queue new URLs on a source-by-source basis. This fixes both the problem where it would find "URLs" in the middle of big areas of binary video content; and where it would encounter a login page and get stuck there.